# lab8/src/static/openapi.json
{
    "swagger": "2.0",
    "info": {
        "title": "Scooter API",
        "description": "API for Scooters and whatever...",
        "version": "1.69.420"
    },
    "paths": {
        "/api/electro_scooters": {
            "get": {
                "summary": "Get all scooters",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No scooters found"
                    }
                }
            },
            "post": {
                "summary": "Create new scooter",
                "parameters": [
                    {
                        "name": "Scooter",
                        "in": "body",
                        "description": "Scooter object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Scooter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Scooter created"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "403": {
                        "description": "Not allowed to add/modify data"
                    }
                }
            },
            "put": {
                "summary": "Update scooter",
                "parameters": [
                    {
                        "name": "Scooter",
                        "in": "body",
                        "description": "Scooter object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Scooter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scooter updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "403": {
                        "description": "Not allowed to add/modify data"
                    },
                    "404": {
                        "description": "Scooter not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete all scooters",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Scooter id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "403": {
                        "description": "Not allowed to add/modify data"
                    },
                    "404": {
                        "description": "Scooter not found"
                    }
                }
            }
        },
        "/api/electro_scooters/{id}": {
            "get": {
                "summary": "Get scooter by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Scooter id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Scooter not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Scooter": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "battery_level": {
                    "type": "number"
                }
            }
        }
    }
}